// app/types/user.ts
export default interface User {
  id: string;
  email: string;
  userName: string;
  profileImg: string;
  banner: string;
  fName: string;
  lName: string;
  bio: string;
  //   friends: Array<string>;
  //   author: string;
  //   title: string;
  //   subject: string;
  //   content: string;
  //   tags: string[];
  //   timestamp: Date;
  // //   status: "friends" | "follower" | "archived";
}